refactor: constant, controller, logic, util 패키지 구조 변경
