+++ src/main/java/calculator/Application.java
@@ -1,9 +1,11 @@
 package calculator;
 
+import calculator.controller.CalculatorController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        CalculatorRun calculatorRun = new CalculatorRun();
-        calculatorRun.run();
+        CalculatorController calculatorController = new CalculatorController();
+        calculatorController.run();
     }
 }
