+++ src/main/java/calculator/logic/CalculatorComputeResult.java
@@ -1,0 +1,4 @@
+package calculator.logic;
+
+public class CalculatorComputeResult {
+}
