+++ src/main/java/calculator/logic/CalculatorInputParser.java
@@ -1,0 +1,4 @@
+package calculator.logic;
+
+public class CalculatorInputParser {
+}
