--- src/main/java/constant/OutputMessage.java
@@ -1,2 +1,0 @@
-package constant;public class OutputMessage {
-}
