--- src/main/java/util/MessageUtil.java
@@ -1,11 +1,0 @@
-package util;
-
-public class MessageUtil {
-    public void printStringForSum() {
-        System.out.println("덧셈할 문자열을 입력해 주세요. ");
-    }
-
-    public void printResult(Integer result) {
-        System.out.printf("결과 : %d", result);
-    }
-}
